git-svn-id: https://cleartoolkit.svn.sourceforge.net/svnroot/cleartoolkit/HTML5/trunk@1347 9db0787f-aee9-4a4d-a695-80f0cfb65743

+++ com.farata.cleardatabuilder.js/resources/java_sample/src/com/farata/java_test/service/IAssociateService.java
@@ -14,5 +14,5 @@ public interface IAssociateService {
 	@CX_JSGenerateSample
 	@CX_JSGenerateStore
 	@CX_JSFillMethod(autoSyncEnabled = true)
-	List<AssociateDTO> getAssociates(Long companyId);
+	List<AssociateDTO> getAssociates(Integer companyId);
 }
